requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
multidict==6.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.0.4
yarl==1.13.1
